create buildindex class and move some funcs to it
